Cypress: Move methods from commands.ts to POs
