--- Dfe.Academies.External.Web/CypressTests/customConfig.ts
@@ -1,4 +1,0 @@
-export const dfeSignInTestEnvURLForA2BDevAndA2BTest: string = 'https://test-interactions.signin.education.gov.uk//7fbd2f4e-8296-4211-a7e4-a38df63d3ff5/usernamepassword'
-export const dfeSignInTestEnvForgotPasswordCodeInputURLForA2BDevAndA2BTest: string = 'https://test-interactions.signin.education.gov.uk/bb11be36-f9b9-420d-8765-aeab083b495d/resetpassword/1786759B-D42E-4A52-B189-4788B76DD793/confirm'
-export const dfeSignInTestEnvCreateAccountForA2BDevAndA2BTest: string = 'https://test-profile.signin.education.gov.uk/register'
-export const unAuthUrl: string = 'https://s184t01-a2bextcdnendpoint-ezfhhdbpembpanh5.z01.azurefd.net/application-overview?appId=10021'
