+++ Dfe.Academies.External.Web/CypressTests/cypress/page-objects/pages/consultationDetails.ts
@@ -1,6 +1,25 @@
 class ConsultationDetails {
   public consultationDetailsElementsVisible(): this {
-    cy.consultationDetailsElementsVisible()
+    cy.get('.govuk-back-link').contains('Back')
+
+    cy.get('.govuk-caption-l').contains('Plymstock School')
+
+    cy.get('.govuk-heading-l').contains('Consultation')
+
+    cy.get('.govuk-body').eq(0).contains('Schools must consult any stakeholders relevant to the conversion')
+
+    cy.get('#role-hint').contains('Has the governing body consulted the relevant stakeholders?')
+
+    cy.get('#consultationStakeholdersOptionYes').should('not.be.checked')
+    cy.get('label[for=consultationStakeholdersOptionYes]').contains('Yes')
+
+    cy.get('#consultationStakeholdersOptionNo').should('not.be.checked')
+    cy.get('label[for=consultationStakeholdersOptionNo]').contains('No')
+
+    cy.get('label[for=SchoolConsultationStakeholdersConsult]').contains('When does the governing body plan to consult?')
+    cy.get('#SchoolConsultationStakeholdersConsult').should('be.enabled')
+
+    cy.get('input[type=submit]').should('be.visible').contains('Save and return to overview')
 
     return this
   }
@@ -13,13 +32,14 @@ class ConsultationDetails {
   }
 
   public fillConsultationDetails(): this {
-    cy.fillConsultationDetails()
+    const consultationDetails = 'When does the governing body plan to consult?'
+    cy.get('#SchoolConsultationStakeholdersConsult').type(consultationDetails)
 
     return this
   }
 
   public submitConsultationDetails(): this {
-    cy.submitConsultationDetails()
+    cy.get('input[type=submit]').click()
 
     return this
   }
