--- Dfe.Academies.External.Web/CypressTests/cypress/page-objects/pages/createAnAccount.ts
@@ -1,23 +1,0 @@
-class CreateAnAccount {
-  public createAccountElementsVisible(): this {
-    cy.createAccountElementsVisible()
-
-    return this
-  }
-
-  public createAccountFailsWithNoData(): this {
-    cy.createAccountFailsWithNoData()
-
-    return this
-  }
-
-  public createAccountSuccessful(): this {
-    cy.createAccountSuccessful()
-
-    return this
-  }
-}
-
-const createAnAccount = new CreateAnAccount()
-
-export default createAnAccount
