--- Dfe.Academies.External.Web/CypressTests/cypress/page-objects/pages/createAnAccountConfirm.ts
@@ -1,11 +1,0 @@
-class CreateAnAccountConfirm {
-  public createAccountConfirmElementsVisible(): this {
-    cy.createAccountConfirmElementsVisible()
-
-    return this
-  }
-}
-
-const createAnAccountConfirm = new CreateAnAccountConfirm()
-
-export default createAnAccountConfirm
