+++ Dfe.Academies.External.Web/CypressTests/cypress/page-objects/pages/declarationSummary.ts
@@ -1,6 +1,17 @@
 class DeclarationSummary {
   public declarationSummaryElementsVisible(): this {
-    cy.declarationSummaryElementsVisible()
+    cy.get('.govuk-back-link').contains('Back')
+
+    cy.get('.govuk-caption-l').contains('Plymstock School')
+
+    cy.get('.govuk-heading-l').contains('Declaration')
+
+    cy.get('a[class="govuk-button govuk-button--secondary"]').should('be.visible').contains('Start section')
+
+    cy.get('b').eq(0).contains('I agree with all of these statements, and believe that the facts stated in this application are true')
+    cy.get('p').eq(2).contains('You have not added any information')
+
+    cy.get('.govuk-button').should('be.visible').contains('Back')
 
     return this
   }
@@ -12,7 +23,18 @@ class DeclarationSummary {
   }
 
   public declarationSummaryCompleteElementsVisible(): this {
-    cy.declarationSummaryCompleteElementsVisible()
+    cy.get('.govuk-back-link').contains('Back')
+
+    cy.get('.govuk-caption-l').contains('Plymstock School')
+
+    cy.get('.govuk-heading-l').contains('Declaration')
+
+    cy.get('.govuk-link').contains('Change your answers')
+
+    cy.get('b').eq(0).contains('I agree with all of these statements, and believe that the facts stated in this application are true')
+    cy.get('p').eq(2).contains('Yes')
+
+    cy.get('.govuk-button').should('be.visible').contains('Back')
 
     return this
   }
