+++ Dfe.Academies.External.Web/CypressTests/cypress/page-objects/pages/fam/trustOverview.ts
@@ -1,8 +1,22 @@
-// TODO pull out methods from commands.ts
-
 class TrustOverview {
   public FAMTrustOverviewTrustNameCompleteElementsVisible(): this {
-    cy.FAMTrustOverviewTrustNameCompleteElementsVisible()
+    cy.get('.govuk-back-link').contains('Back')
+
+    cy.get('.govuk-caption-l').contains('Form a new multi-academy trust')
+
+    cy.get('.govuk-heading-l').contains('Plymouth')
+
+    cy.get('.govuk-body').contains('Complete all sections.')
+
+    cy.get('strong').eq(1).contains('Completed')
+    cy.get('strong').eq(2).contains('Not Started')
+    cy.get('strong').eq(3).contains('Not Started')
+    cy.get('strong').eq(4).contains('Not Started')
+    cy.get('strong').eq(5).contains('Not Started')
+    cy.get('strong').eq(6).contains('Not Started')
+    cy.get('strong').eq(7).contains('Not Started')
+
+    cy.get('.govuk-button').should('be.visible').contains('Return to your application')
 
     return this
   }
@@ -14,7 +28,23 @@ class TrustOverview {
   }
 
   public FAMTrustOverviewOpeningDateCompleteElementsVisible(): this {
-    cy.FAMTrustOverviewOpeningDateCompleteElementsVisible()
+    cy.get('.govuk-back-link').contains('Back')
+
+    cy.get('.govuk-caption-l').contains('Form a new multi-academy trust')
+
+    cy.get('.govuk-heading-l').contains('Plymouth')
+
+    cy.get('.govuk-body').contains('Complete all sections.')
+
+    cy.get('strong').eq(1).contains('Completed')
+    cy.get('strong').eq(2).contains('Completed')
+    cy.get('strong').eq(3).contains('Not Started')
+    cy.get('strong').eq(4).contains('Not Started')
+    cy.get('strong').eq(5).contains('Not Started')
+    cy.get('strong').eq(6).contains('Not Started')
+    cy.get('strong').eq(7).contains('Not Started')
+
+    cy.get('.govuk-button').should('be.visible').contains('Return to your application')
 
     return this
   }
@@ -26,7 +56,23 @@ class TrustOverview {
   }
 
   public FAMTrustOverviewReasonsForFormingTrustCompleteElementsVisible(): this {
-    cy.FAMTrustOverviewReasonsForFormingTrustCompleteElementsVisible()
+    cy.get('.govuk-back-link').contains('Back')
+
+    cy.get('.govuk-caption-l').contains('Form a new multi-academy trust')
+
+    cy.get('.govuk-heading-l').contains('Plymouth')
+
+    cy.get('.govuk-body').contains('Complete all sections.')
+
+    cy.get('strong').eq(1).contains('Completed')
+    cy.get('strong').eq(2).contains('Completed')
+    cy.get('strong').eq(3).contains('Completed')
+    cy.get('strong').eq(4).contains('Not Started')
+    cy.get('strong').eq(5).contains('Not Started')
+    cy.get('strong').eq(6).contains('Not Started')
+    cy.get('strong').eq(7).contains('Not Started')
+
+    cy.get('.govuk-button').should('be.visible').contains('Return to your application')
 
     return this
   }
@@ -38,7 +84,23 @@ class TrustOverview {
   }
 
   public FAMTrustOverviewPlansForGrowthCompleteElementsVisible(): this {
-    cy.FAMTrustOverviewPlansForGrowthCompleteElementsVisible()
+    cy.get('.govuk-back-link').contains('Back')
+
+    cy.get('.govuk-caption-l').contains('Form a new multi-academy trust')
+
+    cy.get('.govuk-heading-l').contains('Plymouth')
+
+    cy.get('.govuk-body').contains('Complete all sections.')
+
+    cy.get('strong').eq(1).contains('Completed')
+    cy.get('strong').eq(2).contains('Completed')
+    cy.get('strong').eq(3).contains('Completed')
+    cy.get('strong').eq(4).contains('Completed')
+    cy.get('strong').eq(5).contains('Not Started')
+    cy.get('strong').eq(6).contains('Not Started')
+    cy.get('strong').eq(7).contains('Not Started')
+
+    cy.get('.govuk-button').should('be.visible').contains('Return to your application')
 
     return this
   }
@@ -50,7 +112,23 @@ class TrustOverview {
   }
 
   public FAMTrustOverviewSchoolImprovementStrategyCompleteElementsVisible(): this {
-    cy.FAMTrustOverviewSchoolImprovementStrategyCompleteElementsVisible()
+    cy.get('.govuk-back-link').contains('Back')
+
+    cy.get('.govuk-caption-l').contains('Form a new multi-academy trust')
+
+    cy.get('.govuk-heading-l').contains('Plymouth')
+
+    cy.get('.govuk-body').contains('Complete all sections.')
+
+    cy.get('strong').eq(1).contains('Completed')
+    cy.get('strong').eq(2).contains('Completed')
+    cy.get('strong').eq(3).contains('Completed')
+    cy.get('strong').eq(4).contains('Completed')
+    cy.get('strong').eq(5).contains('Completed')
+    cy.get('strong').eq(6).contains('Not Started')
+    cy.get('strong').eq(7).contains('Not Started')
+
+    cy.get('.govuk-button').should('be.visible').contains('Return to your application')
 
     return this
   }
@@ -62,7 +140,23 @@ class TrustOverview {
   }
 
   public FAMTrustOverviewGovernanceStructureCompleteElementsVisible(): this {
-    cy.FAMTrustOverviewGovernanceStructureCompleteElementsVisible()
+    cy.get('.govuk-back-link').contains('Back')
+
+    cy.get('.govuk-caption-l').contains('Form a new multi-academy trust')
+
+    cy.get('.govuk-heading-l').contains('Plymouth')
+
+    cy.get('.govuk-body').contains('Complete all sections.')
+
+    cy.get('strong').eq(1).contains('Completed')
+    cy.get('strong').eq(2).contains('Completed')
+    cy.get('strong').eq(3).contains('Completed')
+    cy.get('strong').eq(4).contains('Completed')
+    cy.get('strong').eq(5).contains('Completed')
+    cy.get('strong').eq(6).contains('Completed')
+    cy.get('strong').eq(7).contains('Not Started')
+
+    cy.get('.govuk-button').should('be.visible').contains('Return to your application')
 
     return this
   }
@@ -74,7 +168,23 @@ class TrustOverview {
   }
 
   public FAMTrustOverviewKeyPeopleCompleteElementsVisible(): this {
-    cy.FAMTrustOverviewKeyPeopleCompleteElementsVisible()
+    cy.get('.govuk-back-link').contains('Back')
+
+    cy.get('.govuk-caption-l').contains('Form a new multi-academy trust')
+
+    cy.get('.govuk-heading-l').contains('Plymouth')
+
+    cy.get('.govuk-body').contains('Complete all sections.')
+
+    cy.get('strong').eq(1).contains('Completed')
+    cy.get('strong').eq(2).contains('Completed')
+    cy.get('strong').eq(3).contains('Completed')
+    cy.get('strong').eq(4).contains('Completed')
+    cy.get('strong').eq(5).contains('Completed')
+    cy.get('strong').eq(6).contains('Completed')
+    cy.get('strong').eq(7).contains('Completed')
+
+    cy.get('.govuk-button').should('be.visible').contains('Return to your application')
 
     return this
   }
