--- Dfe.Academies.External.Web/CypressTests/cypress/page-objects/pages/forgottenYourPasswordVerifyCode.ts
@@ -1,11 +1,0 @@
-class ForgottenYourPasswordVerifyCode {
-  public forgotPasswordVerifyCodeElementsVisible(): this {
-    cy.forgotPasswordVerifyCodeElementsVisible()
-
-    return this
-  }
-}
-
-const forgottenYourPasswordVerifyCode = new ForgottenYourPasswordVerifyCode()
-
-export default forgottenYourPasswordVerifyCode
