+++ Dfe.Academies.External.Web/CypressTests/cypress/page-objects/pages/loansSummary.ts
@@ -1,6 +1,19 @@
 class LoansSummary {
   public loansSummaryElementsVisible(): this {
-    cy.loansSummaryElementsVisible()
+    cy.get('.govuk-back-link').contains('Back')
+
+    cy.get('.govuk-caption-l').contains('Finances (Step 4 of 6)')
+    cy.get('.govuk-heading-l').contains('Loans')
+
+    cy.get('legend').contains('Are there any existing loans?')
+
+    cy.get('#anyLoansOptionYes').should('not.be.checked')
+    cy.get('label[for=anyLoansOptionYes]').contains('Yes')
+
+    cy.get('#anyLoansOptionNo').should('not.be.checked')
+    cy.get('label[for=anyLoansOptionNo]').contains('No')
+
+    cy.get('input[type=submit]').should('be.visible').contains('Continue')
 
     return this
   }
@@ -13,7 +26,7 @@ class LoansSummary {
   }
 
   public submitLoansSummary(): this {
-    cy.submitLoansSummary()
+    cy.get('input[type=submit]').click()
 
     return this
   }
