+++ Dfe.Academies.External.Web/CypressTests/cypress/page-objects/pages/preOpeningSupportGrantSummary.ts
@@ -1,6 +1,17 @@
 class PreOpeningSupportGrantSummary {
   public preopeningSupportGrantSummaryElementsVisible(): this {
-    cy.preopeningSupportGrantSummaryElementsVisible()
+    cy.get('.govuk-back-link').contains('Back')
+
+    cy.get('.govuk-caption-l').contains('Plymstock School')
+
+    cy.get('.govuk-heading-l').contains('Pre-opening support grant')
+
+    cy.get('a[class="govuk-button govuk-button--secondary"]').should('be.visible').contains('Start section')
+
+    cy.get('b').eq(0).contains('Do you want these funds paid to the school or the trust the school is joining?')
+    cy.get('p').eq(2).contains('You have not added any information')
+
+    cy.get('.govuk-button').should('be.visible').contains('Back to application overview')
 
     return this
   }
@@ -12,7 +23,18 @@ class PreOpeningSupportGrantSummary {
   }
 
   public preopeningSupportGrantSummaryCompleteElementsVisible(): this {
-    cy.preopeningSupportGrantSummaryCompleteElementsVisible()
+    cy.get('.govuk-back-link').contains('Back')
+
+    cy.get('.govuk-caption-l').contains('Plymstock School')
+
+    cy.get('.govuk-heading-l').contains('Pre-opening support grant')
+
+    cy.get('.govuk-link').contains('Change your answers')
+
+    cy.get('b').eq(0).contains('Do you want these funds paid to the school or the trust the school is joining?')
+    cy.get('p').eq(2).contains('To the school')
+
+    cy.get('.govuk-button').should('be.visible').contains('Back')
 
     return this
   }
