+++ Dfe.Academies.External.Web/CypressTests/cypress/page-objects/pages/whatIsYourRole.ts
@@ -1,6 +1,12 @@
 class WhatIsYourRole {
   public whatIsYourRoleElementsVisible(): this {
-    cy.whatIsYourRoleElementsVisible()
+    cy.get('h1').contains('What is your role?')
+    cy.get('p').contains('Anyone from the school can contribute to this form, but only the chair of governors can complete the declaration section or submit it.')
+    cy.get('input[type=radio]').eq(0).should('exist')
+    cy.get('label[for=RoleTypeChairOfGovernors]').contains('The chair of the school\'s governors')
+    cy.get('input[type=radio]').eq(1).should('exist')
+    cy.get('label[for=RoleTypeOther]').contains('Something else')
+    cy.get('input[type=submit]').should('be.visible').contains('Save and continue')
 
     return this
   }
@@ -9,18 +15,6 @@ class WhatIsYourRole {
     cy.get('input[type=radio]').eq(0).click()
     cy.get('input[type=radio]').eq(0).should('be.checked')
     cy.get('input[type=submit]').click()
-
-    return this
-  }
-
-  public selectSomethingElseRadioButton(): this {
-    cy.selectSomethingElseRadioButton()
-
-    return this
-  }
-
-  public verifySomethingElseRadioButtonChecked(): this {
-    cy.verifySomethingElseRadioButtonChecked()
 
     return this
   }
